requirements.txt
Flask==2.3.0
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.2.0
groq==0.4.0
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
